filter-service/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.8.4
